Create Prime Numbers
Instructions from your teacher:
Write a Python function that takes a number as a parameter and check the number is prime or not.  Find all the prime numbers from 1 to 100.

Note : A prime number (or a prime) is a natural number greater than 1 and that has no positive divisors other than 1 and itself.