Create Numbers
Instructions from your teacher:
Pythagoras and other philosophers of the time believed that because 
mathematical concepts were more "practical" (easier to regulate and 
classify) than physical ones, they had greater actuality. 
St. Augustine of Hippo (A.D. 354–430), wrote: "Numbers are the 
Universal language offered by the deity to humans as confirmation of 
the truth.

Write a program that will decipher your birthday and produce a description 
of the type of person you are.  To do so use the following algorithm and data.

Birthday:  March 15, 1969
3
1+5
1+9+5+7

3+6+22=31
Now, if the number is 11 or 22 then we have a prediction, otherwise,
we are trying to reduce the date to a single digit therefore:

3+1=4

Using 4, query the database for an answer.
Format the output.