Create Tkinter Calculator
Instructions from your teacher:
Build a Tkinter calculator.  It should be able to add, subtract, multiply and divide.  You should be able to clear the screen.

The program will consist of 10 buttons for numbers, one button for clear, a small menu for exiting the program, and a text box for input and output.

DO NOT COPY/PASTE CODE FROM THE WEB:  INSTANT F